runtime: extract weave version execution
- move prepared weave execution and version planning/write helpers out of the runtime weave façade
- add shared execution config, workspace path, and progress helper modules
- keep public runtime weave exports stable while reducing weave.ts to the command façade plus page generation
- update runtime codebase overview
