+++ src/runtime/weave/progress.ts
@@ -1,0 +1,8 @@
+export interface WeaveProgressEvent {
+  designatorPath: string;
+  completed: number;
+  total: number;
+  percent: number;
+}
+
+export type WeaveProgressHandler = (event: WeaveProgressEvent) => void;
